Add static descriptor set layouts in model classes

+++ Earth/Timer.cpp
@@ -8,15 +8,15 @@ Timer::~Timer()
 {
 }
 
-float Timer::getDeltaSec()
+double Timer::getDeltaSec()
 {
-	float deltaSec = 0;
+	double deltaSec = 0;
 
 	time_point<steady_clock> now = high_resolution_clock::now();
 
 	if (lastTimePoint != time_point<steady_clock>::max())
 	{
-		deltaSec = duration_cast<milliseconds>(now - lastTimePoint).count() / 1000.0f;
+		deltaSec = duration_cast<milliseconds>(now - lastTimePoint).count() / 1000.0;
 	}
 
 	lastTimePoint = now;
